refactor: extract faculty dashboard controllers
